+++ src/pages/_app.tsx
@@ -1,7 +1,12 @@
 import "tailwindcss/tailwind.css";
 import type { AppProps } from "next/app";
+import Layout from "../components/Layout";
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  );
 }
 export default MyApp;
